+++ main.go
@@ -16,7 +16,7 @@ type config struct {
 	ListenAddr      string        `env:"LISTEN_ADDR, default=:8080"`
 	ReadTimeout     time.Duration `env:"READ_TIMEOUT, default=30s"`
 	WriteTimeout    time.Duration `env:"WRITE_TIMEOUT, default=30s"`
-	ShutdownTimeout time.Duration `env:"SHUTDOWN_TIMEOUT, default=5m"`
+	ShutdownTimeout time.Duration `env:"SHUTDOWN_TIMEOUT, default=10s"`
 }
 
 func main() {